refactor tradebox trade ratios
